cont

+++ src/sdl.h
@@ -214,7 +214,7 @@ sdl_sensor_info_t *sdl_sensor_get_info_tbl(int *max);
 int sdl_sensor_find(int type);  // returns sensor id, or -1 if not found
 int sdl_sensor_read_raw(int id, double *data, int num_values);
 
-int sdl_sensor_read_step_counter(unsigned long *step_count);
+int sdl_sensor_read_step_counter(double *step_count);
 int sdl_sensor_read_mag_heading(double *mag_heading);
 int sdl_sensor_read_tilt(double *roll, double *pitch);
 int sdl_sensor_read_pressure(double *millibars);
